BABEL: Skip locking the tuple for ENR-related tuple modifications.
Commit 09f0820095ea8b7c6ca4269454a94695c8421628 introduced
locking for individual tuples during certain operations. However,
Babelfish introduced the concept of ENR-only relations, which
store all catalog tuples in their own local cache and not in the
physical catalogs. As these relations and their catalog tuples
are completely session-local, there is no need to acquire locks on
tuples for these relations (and it would just lead to a crash
anyways since the tuples do not have an underlying TID to lock against).

Signed-off-by: Jason Teng <jasonten@amazon.com>