Fix description of PostgreSQL::Test::Cluster::wait_for_event()
The arguments of the function were listed in an incorrect order in the
description of the routine.  This information can be seen with perldoc.

Issue spotted while working on this area of the code.

Backpatch-through: 17
(cherry picked from commit b8d08aafc0ef8b5f7bf15dcf4bc6af736ee81f15)